// src/enum/auth.message.enum.ts
export enum AUTH_MESSAGE {
    FETCHED_SUCCESS = 'User fetched successfully.',
    CREATED_SUCCESS = 'User created successfully.',
    UPDATED_SUCCESS = 'User updated successfully.',
    DELETED_SUCCESS = 'User deleted successfully.',
    NOT_FOUND = 'User not found.',
    INVALID_INPUT = 'Invalid input provided.'
}